:sparkles: feat(frontend): useLottieLike 좋아요 커스텀훅 추가 (#64)
